% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_auxiliares.R
\name{buscar_por_titulo}
\alias{buscar_por_titulo}
\title{Buscar obras por palavra-chave no título}
\usage{
buscar_por_titulo(palavra)
}
\arguments{
\item{palavra}{Palavra-chave para busca}
}
\value{
Data frame com obras que contêm a palavra no título
}
\description{
Buscar obras por palavra-chave no título
}
\examples{
buscar_por_titulo("Casmurro")
}
