% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_auxiliares.R
\name{filtrar_por_genero}
\alias{filtrar_por_genero}
\title{Filtrar obras por gênero}
\usage{
filtrar_por_genero(genero)
}
\arguments{
\item{genero}{Vetor de gêneros literários}
}
\value{
Data frame com metadados das obras filtradas
}
\description{
Filtrar obras por gênero
}
\examples{
filtrar_por_genero("romance")
filtrar_por_genero(c("romance", "conto"))
}
