% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obras_metadata}
\alias{obras_metadata}
\title{Metadados das obras de Machado de Assis}
\format{
Um data frame com colunas:
\describe{
\item{titulo}{Título da obra}
\item{ano}{Ano de publicação (quando disponível)}
\item{genero}{Gênero literário (romance, conto, poesia, crônica, teatro)}
\item{arquivo}{Nome do arquivo PDF original}
}
}
\source{
\url{https://machado.mec.gov.br/}
}
\usage{
obras_metadata
}
\description{
Dataset contendo informações sobre as obras completas de Machado de Assis
disponíveis no site do MEC.
}
\keyword{datasets}
