% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_auxiliares.R
\name{obter_texto}
\alias{obter_texto}
\title{Obter texto de uma obra específica}
\usage{
obter_texto(titulo)
}
\arguments{
\item{titulo}{Título da obra}
}
\value{
Texto completo da obra
}
\description{
Obter texto de uma obra específica
}
\examples{
obter_texto("Dom Casmurro")
}
