% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_auxiliares.R
\name{preparar_tidytext}
\alias{preparar_tidytext}
\title{Preparar texto para análise com tidytext}
\usage{
preparar_tidytext(titulo)
}
\arguments{
\item{titulo}{Título da obra}
}
\value{
Data frame tokenizado pronto para tidytext
}
\description{
Preparar texto para análise com tidytext
}
\examples{
preparar_tidytext("Memórias Póstumas de Brás Cubas")
}
