% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stopwords_pt}
\alias{stopwords_pt}
\title{Stopwords em Português}
\format{
Um tibble com uma coluna:
\describe{
\item{palavra}{A palavra vazia (character)}
}
}
\usage{
stopwords_pt
}
\description{
Lista de palavras vazias (stopwords) em português para filtragem
em análises de texto.
}
\keyword{datasets}
